자동저장!!

+++ src/content/main.php
@@ -23,7 +23,7 @@ function printContentPc(){
 		</table>
         <div class="grid-stack">
             <?php
-                //if(!file_exists("data/user/main_layout/{$me['n_id']}.txt")) {
+                if(!file_exists("data/user/main_layout/{$me['n_id']}.txt")) {
                     file_put_contents("data/user/main_layout/{$me['n_id']}.txt", <<<JSON
                     [
                        {
@@ -148,7 +148,7 @@ function printContentPc(){
                     ]
 JSON
                                     );
-                //}
+                }
                 $modules = json_decode(file_get_contents("data/user/main_layout/{$me['n_id']}.txt"), true);
                 allModules($modules);
             ?>
